api/agente-ia/requirements.txt
google-adk
uvicorn
fastapi
python-dotenv
requests 